requirements.txt
requests
selenium
Pillow
PyPDF2